+++ app/Basket/Synchronisation/InstallationSynchronisationService.php
@@ -47,10 +47,9 @@ class InstallationSynchronisationService extends AbstractSynchronisationService
     {
         $installation = $this->fetchInstallationLocalObject($id);
 
-        $merchant = $this->fetchMerchantLocalObject($installation->merchant_id);
-
         try {
-            $installationEntity = $this->installationGateway->getInstallation($installation->ext_id, $merchant->token);
+            $installationEntity = $this->installationGateway
+                ->getInstallation($installation->ext_id, $installation->merchant->token);
         } catch (\Exception $e) {
 
             $installation->linked = false;
@@ -178,6 +177,7 @@ class InstallationSynchronisationService extends AbstractSynchronisationService
 
         return $rtn;
     }
+
     /**
      * @author WN
      * @param InstallationEntity $installationEntity