+++ resources/views/applications/panels/flexible-finance.blade.php
@@ -1,5 +1,9 @@
 <div role="tabpanel" class="tab-pane active" id="prod-FF">
-    {!! Form::open(['action' => ['InitialisationController@request', $location->id], 'class' => 'initialiseForm']) !!}
+    @if(isset($assisted) && $assisted == true)
+        {!! Form::open(['action' => ['InitialisationController@requestAssisted', $location->id], 'class' => 'initialiseForm']) !!}
+    @else
+        {!! Form::open(['action' => ['InitialisationController@request', $location->id], 'class' => 'initialiseForm']) !!}
+    @endif
     <h2>Flexible Finance</h2>
     <div class="form-group container-fluid">
         <div class="row text-center">
@@ -99,14 +103,15 @@
                 </div>
             </div>
             <div class="row">
-                @if($bitwise->contains(2) && count($bitwise->explode()) > 1)<div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
-                        @if($bitwise->contains(2))
-                            <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin">Continue with In-store Application</button>
-                        @endif
-                    </div>
-                    @if($bitwise->contains(2) && count($bitwise->explode()) > 1) <div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
+                    <div class="row">
+                        @if($bitwise->contains(2) && count($bitwise->explode()) > 1) <div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
+                            @if($bitwise->contains(2))
+                                <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin">Continue with In-store Application</button>
+                            @endif
+                        </div>
+                        @if($bitwise->contains(2) && count($bitwise->explode()) > 1) <div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
                             @if($bitwise->contains(4) || $bitwise->contains(8))
-                                <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin" name="alternate" value="true">Create an Application Link</button>
+                                <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin" name="assisted" value="true">Create an Application Link</button>
                             @endif
                         </div>
                     </div>