+++ resources/views/includes/message/action_response.blade.php
@@ -2,7 +2,7 @@
 
 {{--LARAVEL STANDARD ERRORS--}}
 @if( $errors->any())
-    <div id="actionMessage" hidden="hidden">
+    <div id="actionMessage">
         <div class="alert alert-danger alert-dismissible" role="alert">
             <button type="button" class="close message_close" data-dismiss="alert" aria-label="Close">
                 <span aria-hidden="true">&times;</span>
@@ -16,7 +16,7 @@
 
 @foreach($messages as $k => $v)
     @if($v)
-        <div id="actionMessage" hidden="hidden">
+        <div id="actionMessage">
             <div class="alert alert-{{ ($k == 'error')?'danger':$k }} alert-dismissible" role="alert">
                 <button type="button" class="close message_close" data-dismiss="alert" aria-label="Close">
                     <span aria-hidden="true">&times;</span>