+++ resources/views/initialise/main.blade.php
@@ -46,7 +46,6 @@
             </div>
             {!! Form::close() !!}
         </div>
-
         @if(isset($options) || isset($flexibleFinance))
 
             @if(
@@ -91,9 +90,12 @@
                     @foreach($options as $k => $group)
 
                         @foreach($group['products'] as $l => $product)
-
                             <div role="tabpanel" class="tab-pane{{ ($k == 0 && $l == 0 && (isset($flexibleFinance) && count($flexibleFinance)) == 0)?' active':'' }}" id="prod-{{$product['id']}}">
+                                @if(isset($assisted) && $assisted == true)
+                                    {!! Form::open(['action' => ['InitialisationController@requestAssisted', $location->id], 'class' => 'initialiseForm']) !!}
+                                @else
                                 {!! Form::open(['action' => ['InitialisationController@request', $location->id], 'class' => 'initialiseForm']) !!}
+                                @endif
                                     <h2>{{$product['name']}}</h2>
                                     <div class="form-group container-fluid">
                                         <div class="row text-center">
@@ -306,9 +308,9 @@
                                                 <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin">Continue with In-store Application</button>
                                             @endif
                                         </div>
-                                            @if($bitwise->contains(2) && count($bitwise->explode()) > 1) <div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
+                                        @if($bitwise->contains(2) && count($bitwise->explode()) > 1) <div class="col-sm-6 col-xs-12">@else <div class="col-sm-12 col-xs-12">@endif
                                             @if($bitwise->contains(4) || $bitwise->contains(8))
-                                                <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin" name="alternate" value="true">Create an Application Link</button>
+                                                <button type="submit" class="btn btn-success btn-lg btn-block btn-bottom-margin" name="assisted" value="true">Create an Application Link</button>
                                             @endif
                                         </div>
                                     </div>